Corrección en Crear y Editar Evento

+++ Decimatio.Domain/Interfaces/IEventoRepository.cs
@@ -3,7 +3,7 @@
     public interface IEventoRepository
     {
         Task<IEnumerable<Evento>> GetAllEventos();
-        Task<Evento> GetById(int idEvento);
+        Task<Evento> GetById(long idEvento);
         Task<int> AddEvento(Evento evento);
         Task<bool> UpdateEvento(Evento evento);
         Task<bool> DeleteEvento(int idEvento);
